+++ .gitignore
@@ -1,3 +1,4 @@
 *.log
 node_modules
 coverage*
+dist
