+++ .nyc_output/processinfo/27a49456-986a-4cce-89a1-d19b3c87def7.json
@@ -1,0 +1,1 @@
+{"parent":null,"pid":41026,"argv":["/home/elianiva/.fnm/node-versions/v16.14.0/installation/bin/node","/home/elianiva/Dev/teknologi-umum/httpsnippet/node_modules/mocha/bin/_mocha","--require","ts-node/register","test/**/*.js","--exit"],"execArgv":[],"cwd":"/home/elianiva/Dev/teknologi-umum/httpsnippet","time":1647081536354,"ppid":40989,"coverageFilename":"/home/elianiva/Dev/teknologi-umum/httpsnippet/.nyc_output/27a49456-986a-4cce-89a1-d19b3c87def7.json","externalId":"","uuid":"27a49456-986a-4cce-89a1-d19b3c87def7","files":["/home/elianiva/Dev/teknologi-umum/httpsnippet/src/targets/node/request.js","/home/elianiva/Dev/teknologi-umum/httpsnippet/src/helpers/code-builder.ts"]}
