+++ package.json
@@ -51,40 +51,50 @@
   "scripts": {
     "quick": "mocha --no-timeouts --fgrep 'Request Validation' --invert",
     "pretest": "standard && echint",
-    "test": "mocha --no-timeouts",
+    "test": "mocha -r ts-node/register --no-timeouts",
     "posttest": "exit 0 && npm run coverage",
     "coverage": "istanbul cover --dir coverage _mocha -- --fgrep 'Request Validation' --invert -R dot",
-    "codeclimate": "codeclimate < coverage/lcov.info"
+    "codeclimate": "codeclimate < coverage/lcov.info",
+    "build": "tsc"
   },
   "standard": {
     "env": "mocha",
     "ignore": [
-      "**/test/fixtures/**"
+      "**/test/fixtures/**",
+      "**/dist/**"
     ]
   },
   "echint": {
     "ignore": [
       "coverage/**",
       "**/node_modules/**",
-      "**/fixtures/**"
+      "**/fixtures/**",
+      "**/dist/**"
     ]
   },
   "devDependencies": {
+    "@types/debug": "^4.1.7",
+    "@types/event-stream": "^4.0.0",
+    "@types/har-format": "^1.2.8",
+    "@types/node": "^17.0.21",
     "codeclimate-test-reporter": "^0.5.1",
     "echint": "^4.0.2",
     "glob": "^6.0.1",
     "istanbul": "^0.4.0",
     "mocha": "^8.2.1",
+    "nyc": "^15.1.0",
     "require-directory": "^2.1.1",
     "should": "^13.2.3",
-    "standard": "^16.0.3"
+    "standard": "^16.0.3",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.2"
   },
   "dependencies": {
     "chalk": "^1.1.1",
     "commander": "^2.9.0",
     "debug": "^2.2.0",
     "event-stream": "3.3.4",
-    "form-data": "3.0.0",
+    "form-data": "^4.0.0",
     "fs-readfile-promise": "^2.0.1",
     "fs-writefile-promise": "^1.0.3",
     "har-validator": "^5.0.0",
