+++ src/helpers/reducer.ts
@@ -1,6 +1,4 @@
-'use strict'
-
-module.exports = function (obj, pair) {
+export default function reducer(obj, pair) {
   if (obj[pair.name] === undefined) {
     obj[pair.name] = pair.value
     return obj
