+++ src/targets/csharp/httpclient.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const CodeBuilder = require('../../helpers/code-builder')
+const CodeBuilder = require('../../helpers/code-builder').default
 const helpers = require('../../helpers/headers')
 
 function getDecompressionMethods (source) {
