+++ src/targets/java/nethttp.js
@@ -10,7 +10,7 @@
 
 'use strict'
 
-const CodeBuilder = require('../../helpers/code-builder')
+const CodeBuilder = require('../../helpers/code-builder').default
 
 module.exports = function (source, options) {
   const opts = Object.assign(
