+++ src/targets/javascript/axios.js
@@ -11,7 +11,7 @@
 
 const util = require('util')
 const stringifyObject = require('stringify-object')
-const CodeBuilder = require('../../helpers/code-builder')
+const CodeBuilder = require('../../helpers/code-builder').default
 
 module.exports = function (source, options) {
   const opts = Object.assign({
