+++ src/targets/python/requests.js
@@ -11,7 +11,7 @@
 'use strict'
 
 const util = require('util')
-const CodeBuilder = require('../../helpers/code-builder')
+const CodeBuilder = require('../../helpers/code-builder').default
 const helpers = require('./helpers')
 
 module.exports = function (source, options) {
