+++ src/targets/r/httr.js
@@ -11,7 +11,7 @@
 'use strict'
 
 const util = require('util')
-const CodeBuilder = require('../../helpers/code-builder')
+const CodeBuilder = require('../../helpers/code-builder').default
 
 module.exports = function (source, options) {
   // Start snippet
