+++ src/targets/ruby/native.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const CodeBuilder = require('../../helpers/code-builder')
+const CodeBuilder = require('../../helpers/code-builder').default
 
 module.exports = function (source, options) {
   const code = new CodeBuilder()
