+++ src/targets/shell/curl.js
@@ -13,7 +13,7 @@
 const util = require('util')
 const helpers = require('../../helpers/shell')
 const headerHelpers = require('../../helpers/headers')
-const CodeBuilder = require('../../helpers/code-builder')
+const CodeBuilder = require('../../helpers/code-builder').default
 
 module.exports = function (source, options) {
   const opts = Object.assign({
