+++ src/targets/shell/httpie.js
@@ -12,7 +12,7 @@
 
 const util = require('util')
 const shell = require('../../helpers/shell')
-const CodeBuilder = require('../../helpers/code-builder')
+const CodeBuilder = require('../../helpers/code-builder').default
 
 module.exports = function (source, options) {
   const opts = Object.assign({
