+++ src/types/PostData.ts
@@ -1,0 +1,3 @@
+import type { PostData as HARPostData } from "har-format";
+
+export type PostData = HARPostData & { size: number; }
